Functional test for restore_signals, include signal.h. (#51)
restore_signals=True lacked a functional test.

It was possible, though unlikely in my experience, to build the extension module against a Python interpreter that was not already including <signal.h> via Python.h.  I believe this might have happened on builds that did not enable --with-fpectl?  The Debian CPython 2.7.13 package does not exhibit the problem.

Regardless, testing for signal.h and explicitly including it is always the right thing to do.